Misc. cleanup bits I missed in my last pass.
I apparently forgot to convert an automation message declaration from int to size_t, not sure how everything compiled cleanly everywhere...

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/6465003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74217 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d8ede05564c868d640f56c5b94a21acb080079df

+++ chrome/common/automation_messages_internal.h
@@ -1414,7 +1414,7 @@ IPC_SYNC_MESSAGE_CONTROL2_1(AutomationMsg_WaitForTabCountToBecome,
 // Waits for the infobar count to reach given number.
 IPC_SYNC_MESSAGE_CONTROL2_1(AutomationMsg_WaitForInfoBarCount,
                             int /* tab handle */,
-                            int /* target count */,
+                            size_t /* target count */,
                             bool /* success */)
 
 // Waits for the autocomplete edit to receive focus.
